fix: change cargoWeight to int

+++ src/main/java/acme/entities/aircraft/Aircraft.java
@@ -45,7 +45,7 @@ public class Aircraft extends AbstractEntity {
 	@Mandatory
 	@ValidNumber(min = 2000, max = 50000)
 	@Automapped
-	private Double				cargoWeight;
+	private Integer				cargoWeight;
 
 	@Mandatory
 	@Valid
